lib: split CSV output into dedicated file
Moev all CSV specific logic into a dedicated file.
